allow configuring disk space allocated to each container
By default, Docker is configured on the AWS ECS AMI to use device
manager to allocate 10GB of disk space to each running container. This
uses thin provisioning to allow over committing the storage on the host.
Expose the dm.basesize Docker option so that projects that want to write
out a large amount of temporary files inside their container can do so.