ecs-cluster: cleanup and simplify autodraining lambda
Reduce the number of API calls to AWS's ECS API by adding a filter to
the call to list_container_instances().

Renamed instance_has_running_tasks() to can_terminate() and put the
checks that was in send_heartbeat() inside the other checks with
can_terminate() since it fits well with the rest of the Lambda function.

There was previously a race condition between the check to see if there
are any tasks running on the instance, and when the instance is set to
draining. After the instance is set to draining, lookup the running task
count again to avoid shutting down an instnace where a job was just
started.

Include the full SNS message in all log messages to make correlating
logs easier since the request ID will be logged.

Lambdas embedded in CloudFormation like this have a maximum size limit
of 4096 bytes and this function was previously at 3853 bytes. All of the
changes above, plus shortening some variable names, reduced the overall
code size to 3041 bytes.